webkit-efl/LayoutTests/platform/efl-tizen/fast/dom/scroll-element-to-rect-expected.txt

Scrolled element to rect
FAIL rect.left should be 65. Was 610.
FAIL rect.top should be 139. Was 810.
PASS successfullyParsed is true

TEST COMPLETE
